separate out api and store
