+++ context/app/static/js/components/savedLists/constants.ts
@@ -1,0 +1,7 @@
+export const SAVED_ENTITIES_DEFAULT = {
+  title: 'My Saved Items',
+  description: '',
+  dateSaved: Date.now(),
+  dateLastModified: Date.now(),
+  savedEntities: {},
+};
